+++ compiler/src/test/kotlin/com/r0adkll/kimchi/TestUtils.kt
@@ -4,6 +4,8 @@ package com.r0adkll.kimchi
 
 import com.squareup.kotlinpoet.asClassName
 import kotlin.reflect.KClass
+import kotlin.reflect.KFunction
+import kotlin.reflect.jvm.kotlinFunction
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
 
@@ -20,20 +22,36 @@ fun Class<*>.getHintScope(hintPackage: String): KClass<*>? =
 fun Class<*>.contributedProperties(
   hintPackageName: String,
 ): List<KClass<*>>? {
+  val clazz = topLevelClass(hintPackageName) ?: return null
+
+  return clazz.declaredFields
+    .sortedBy { it.name }
+    .map { field -> field.use { it.get(null) } }
+    .filterIsInstance<KClass<*>>()
+}
+
+fun Class<*>.topLevelFunctions(
+  packageName: String,
+): List<KFunction<*>>? {
+  val clazz = topLevelClass(packageName) ?: return null
+
+  return clazz.declaredMethods
+    .sortedBy { it.name }
+    .mapNotNull { method -> method.kotlinFunction }
+}
+
+private fun Class<*>.topLevelClass(
+  packageName: String,
+): Class<*>? {
   // The capitalize() comes from kotlinc's implicit handling of file names -> class names. It will
   // always, unless otherwise instructed via `@file:JvmName`, capitalize its facade class.
   val className = kotlin.asClassName()
     .simpleName
     .plus("Kt")
 
-  val clazz = try {
-    classLoader.loadClass("$hintPackageName.$className")
+  return try {
+    classLoader.loadClass("$packageName.$className")
   } catch (e: ClassNotFoundException) {
-    return null
+    null
   }
-
-  return clazz.declaredFields
-    .sortedBy { it.name }
-    .map { field -> field.use { it.get(null) } }
-    .filterIsInstance<KClass<*>>()
 }